# swagger.yaml
openapi: 3.0.0

info:
  version: '1.0'
  title: ErrorSwag
  description: 'Create a community of like minded authors to foster inspiration and innovation by leveraging the modern web.'

servers:
  - url: http://localhost:3000/api/v1

tags:
  - name: 'welcome'
    description: 'API for quick api welcome test'
  - name: 'auth'
    description: 'API for login, signup and account verification'
  - name: 'users'
    description: 'API for users'
  - name: 'articles'
    description: 'API for articles' 
  - name : ratings
    description: 'API for ratings of Articles'

paths:
  /:
    get:
      tags:
        - welcome
      summary: Test get request
      responses:
        '200':
          description: Successfull
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response'
  /auth/login:
    post:
      tags:
        - 'auth'
      summary: 'Logs user into the system'
      description: ''
      operationId: 'loginUser'
      produces:
        - 'application/xml'
        - 'application/json'
      parameters:
        - in: 'query'
          name: 'email'
          schema:
            type: 'string'
            required: true
          description: The user's email for login
        - in: 'query'
          name: 'password'
          schema:
            type: 'string'
            required: true
          description: The password for login in clear text
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response'
        401:
          description: 'failed operation'
  /auth/signup:
    post:
      tags:
        - 'auth'
      summary: 'Create a new user into the system'
      description: ''
      operationId: 'signupUser'
      produces:
        - 'application/xml'
        - 'application/json'
      parameters:
        - in: 'query'
          name: 'email'
          description: "The user's email for signup"
          schema:
            type: 'string'
            required: true
        - in: 'query'
          name: 'username'
          schema:
            type: 'string'
            required: true
          description: 'The username for signup'
        - in: 'query'
          name: 'password'
          description: 'The password for signup in clear text'
          schema:
            type: 'string'
            required: true
      responses:
        200:
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response'
        400:
          description: 'failed operation'
  /auth/verify/{token}:
    patch:
      tags:
        - 'auth'
      summary: "Verifies a user's account"
      description: ''
      operationId: 'verifyUser'
      produces:
        - 'application/xml'
        - 'application/json'
      parameters:
        - in: 'path'
          name: 'token'
          description: 'The token sent to the users email.'
          schema:
            type: 'string'
            required: true
      responses:
        200:
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response'
        400:
          description: "failed operation"
  /users:
    get: 
      tags: 
        - "users"
      summary : List All Users
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: The current page
        - in: query
          name: limit
          schema:
            type: integer
          description: The numbers of items to return
      responses:
        200:
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/response"
        400:              
          description: 'failed operation'
  /user/unfollow:
    patch:
      tags:
        - 'users'
      summary: 'Unfollow users on errorSwag'
      description: 'Id of the user to unfollow'
      operationId: 'followUser'
      produces:
        - 'application/xml'
        - 'application/json'
    parameters:
      - name: 'followeeId'
        in: 'body'
        description: 'Id of the user to unfollow'
        schema:
          type: 'number'
          required: true
    responses:
      200:
        description: 'successful operation'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/response'
    400:
      description: 'Invalid user id'
  /user/follow:
    patch:
      tags:
        - 'users'
      summary: 'Follow other users on errorSwag'
      description: 'Id of the user to follow should be presented as a request body'
      operationId: 'followUser'
      produces:
        - 'application/xml'
        - 'application/json'
    parameters:
      - name: 'followeeId'
        in: 'body'
        description: 'Id of the user to follow'
        schema:
          type: 'number'
          required: true
    responses:
      200:
        description: 'successful operation'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/response'
    400:
      description: 'Invalid user id'
  /users/followers:
    get:
      tags:
        - 'followers'
      summary: 'Get all users that follow this user'
      description: 'Returns all the followers of this user'
      operationId: 'get all followers'
      produces:
        - 'application/xml'
        - 'application/json'
      parameters: []
      responses:
        200:
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response'
        400:
          description: 'failed operation'
  /user/bookmark:
    patch:
      tags:
        - 'bookmark'
      summary: 'Bookmark articles on errorswag'
      description: 'Id of the article to be bookmarked should be presented as a request body'
      operationId: 'bookmark Article'
      produces:
        - 'application/xml'
        - 'application/json'
    parameters:
      - name: 'articleId'
        in: 'body'
        description: 'Id of the article to be bookmarked'
        schema:
          type: 'number'
          required: true
    responses:
      200:
        description: 'successful operation'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/response'
    400:
      description: 'Invalid article id'
  /user/unbookmark:
    patch:
      tags:
        - 'bookmark'
      summary: 'Unbookmark articles on errorSwag'
      description: 'Id of the article to be unbookmarked should be presented as a request body'
      operationId: 'articeId'
      produces:
        - 'application/xml'
        - 'application/json'
    parameters:
      - name: 'articleId'
        in: 'body'
        description: 'Id of the article to unbookmark'
        schema:
          type: 'number'
          required: true
    responses:
      200:
        description: 'successful operation'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/response'
    400:
      description: 'Invalid user id'
  /articles:
    get: 
      tags: 
        - 'articles'
      summary : List All Articles
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: The current page
        - in: query
          name: limit
          schema:
            type: integer
          description: The numbers of items to return
      responses:
        200:
          description: 'Successfull'
          content:
              application/json:
                schema:
                  $ref: "#/components/schemas/response"      
        400:          
          description: 'failed operation'  

  /articles/{articleId}/ratings:
    patch: 
      tags: 
        - ratings
      summary : Rating of Articles
      description: ""
      operationId: "ratingArticles"
      produces:
        - 'application/xml'
        - 'application/json'
      parameters:
        - in: header
          name: x-access-token
          schema:
            type: string
          required: true
          description: token of user
        - in: path
          name: articleId
          schema:
            type: integer
          required: true
          description: articles Id for article been rated
        - in: body
          name: ratings
          schema:
            type: integer
          required: true
          description: ratings for the article
      responses:
        200:
          description: Successfull
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/goodResponse"    
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/badResponse" 
        500:
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/badResponse" 
          description: 'failed operation'

  /articles/:articleId/ratings:
    patch:
      tags:
        - 'ratings'
      summary: 'Get all article ratings'
      description: 'Id of the article to fetch its ratings'
      operationId: 'bookmark Article'
      produces:
        - 'application/xml'
        - 'application/json'
    parameters:
      - name: 'articleId'
        in: 'params'
        description: 'Fetch the ratings of an article'
        schema:
          type: 'number'
          required: true
    responses:
      200:
        description: 'successful operation'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/response'
    400:
      description: 'Invalid article id'
    500:
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/badResponse" 
          description: 'failed operation'
components:
  schemas:
    response:
      type: object
      properties:
        status:
          type: string
        data:
          type: string
    goodResponse:
      type: object
      properties:
        status:
          type: string
        data:
          type: object
    badResponse:
      type: object
      properties:
        message:
          type: string